backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.3",
    "django>=6.0.4",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.3.3",
    "redis>=7.4.0",
]
